Rfactored ServiceConfigurationResolver
